requirements.txt
alembic==1.0.0
aniso8601==3.0.2
appnope==0.1.0
asn1crypto==0.24.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.3.1
cycler==0.10.0
decorator==4.3.0
dominate==2.3.1
entrypoints==0.2.3
enum34==1.1.6
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
functools32==3.2.3.post2
futures==3.2.0
gpio==0.2.0
html5lib==1.0.1
idna==2.7
imutils==0.4.6
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.8.3
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.6.0
numpy==1.14.4
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
pyparsing==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.0
requests==2.19.1
scandir==1.9.0
schema==0.6.8
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.10
subprocess32==3.5.2
sympy==1.2
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
typing==3.5.3.0
urllib3==1.23
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
WTForms==2.2.1


git init
rm -f .git/index.lock
git add .
git commit -m "Initial commit"
git remote set-url origin https://github.com/Eric-Muthemba/futurenet_interview.git
git push -f origin master